backend: change environment management
Secrets are stored in Google Cloud Secret Manager. When using Firebase
emulators, they can be retrieved from a `.secret.local` file.

The `.env.local` file overrides `.env` when using the emulators.  This
will avoid mistakes when deploying to Cloud Functions, as Google Cloud
does not take `.env.local` into account. As a result, the `.env` becomes
the configuration of the prod environment, and is commited.