frontend: load StoryForm from Firestore
The StoryForm provider is removed. Instead, the current StoryForm is put
into CreateStoryState, and initialized with a dedicated method.

This is consistent with the way the state is used: every time we want to
create a new story, we fetch the relevant story form. However, we do not
want this to happen during a story.