backend: refactor StoryForm.getAllFormResponses
Rename it to `getAllAnswers`, since "answer" is the new name to use.

Remove `firestore_story_forms.test.ts`, as it was just repeating the
same test as `story_form.test.ts`.

This commit breaks the implementation of `StoryCacheV1Manager`, this is
intentional and will be repaired when the latter is refactored.